Update add_mastery_check()
The function shouldn't be called during the file, this is because we're going to import the file onto main.py, and use it from there
I changed it so add_mastery_check() will add a new mastery check itself. Instead of taking the name of a mastery check, returning it, and having a json reader dump the new information from there.